Rearrange types and plain data converters
