+++ packages/liveblocks-core/src/comments/__tests__/comment-body.test.ts
@@ -1,9 +1,9 @@
+import type { CommentBody } from "../../types/CommentBody";
 import type { CommentBodyResolveUsersArgs } from "../comment-body";
 import {
   getMentionedIdsFromCommentBody,
   stringifyCommentBody,
 } from "../comment-body";
-import type { CommentBody } from "../types/CommentBody";
 
 function capitalize(string: string) {
   return string.charAt(0).toUpperCase() + string.slice(1);
