+++ packages/liveblocks-core/src/types/BaseMetadata.ts
