+++ packages/liveblocks-core/src/types/InboxNotificationData.ts
